Added AudioValidationBadge component and integrate into AudioWaveformWithTranscription and TextCellEditor
- Introduced the `AudioValidationBadge` component to handle audio validation status and user interactions.
- Refactored `AudioWaveformWithTranscription` to utilize the new badge, removing redundant validation logic and popover handling.
- Updated `TextCellEditor` to include the `AudioValidationBadge`, enhancing audio validation visibility and user experience.
- Streamlined validation handling by centralizing logic within the new component, improving code maintainability.
